menu-commit
Added a pause menu which can be activated by the esc or space keys.  From this menu you can choose to continue the level, restart the level, or quit the game.

Also added a portal object at the end of each level that will load the next scene of the game.